requirements.txt
asgiref==3.8.1
Django==5.2.3
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-js-asset==3.1.2
djangorestframework==3.16.0
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
